+++ colander/__init__.py
@@ -1397,7 +1397,7 @@ class Number(SchemaType):
     num = None
 
     def serialize(self, node, appstruct):
-        if appstruct is null:
+        if appstruct in (null, None):
             return null
 
         try: