+++ colander/tests/test_colander.py
@@ -1750,6 +1750,14 @@ class TestInteger(unittest.TestCase):
         result = typ.serialize(node, val)
         self.assertEqual(result, colander.null)
 
+    def test_serialize_none(self):
+        import colander
+
+        node = DummySchemaNode(None)
+        typ = self._makeOne()
+        result = typ.serialize(node, None)
+        self.assertEqual(result, colander.null)
+
     def test_deserialize_emptystring(self):
         import colander
 
@@ -1810,6 +1818,21 @@ class TestFloat(unittest.TestCase):
         result = typ.serialize(node, val)
         self.assertEqual(result, colander.null)
 
+    def test_serialize_none(self):
+        import colander
+
+        node = DummySchemaNode(None)
+        typ = self._makeOne()
+        result = typ.serialize(node, None)
+        self.assertEqual(result, colander.null)
+
+    def test_serialize_zero(self):
+        val = 0
+        node = DummySchemaNode(None)
+        typ = self._makeOne()
+        result = typ.serialize(node, val)
+        self.assertEqual(result, '0.0')
+
     def test_serialize_emptystring(self):
         import colander
 
@@ -1862,6 +1885,21 @@ class TestDecimal(unittest.TestCase):
         typ = self._makeOne()
         result = typ.serialize(node, val)
         self.assertEqual(result, colander.null)
+
+    def test_serialize_none(self):
+        import colander
+
+        node = DummySchemaNode(None)
+        typ = self._makeOne()
+        result = typ.serialize(node, None)
+        self.assertEqual(result, colander.null)
+
+    def test_serialize_zero(self):
+        val = 0
+        node = DummySchemaNode(None)
+        typ = self._makeOne()
+        result = typ.serialize(node, val)
+        self.assertEqual(result, '0')
 
     def test_serialize_emptystring(self):
         import colander